API Documentation

AHMI — Module

AHMI

The Adaptive Harmonic Mean Integration (AHMI) algorithm.

AHMI.AHMIntegration — Type

struct AHMIntegration <: IntegrationAlgorithm

Adaptive Harmonic Mean Integration algorithm (Caldwell et al.).

Constructors:

AHMIntegration(; fields...)

Fields:

whitening::BAT.WhiteningAlgorithm

Default: CholeskyPartialWhitening()

autocorlen::BAT.AutocorLenAlgorithm

Default: GeyerAutocorLen()

volumetype::Symbol

Default: :HyperRectangle

max_startingIDs::Int64

Default: 10000

max_startingIDs_fraction::Float64

Default: 2.5

rect_increase::Float64

Default: 0.1

warning_minstartingids::Int64

Default: 16

dotrimming::Bool

Default: true

uncertainty::Vector{Symbol}

List of uncertainty estimation methods to use, first entry will be used for primary result. Valid values:

:cov: Integral uncertainty for integration regions is estimated based on covariance of integrals of subsets of samples in the regions

:ess: Integral uncertainty for integration regions is estimated based on estimated effective number of samples in each region.

Default: [:cov]